spacy-tokens/requirements.txt
HTMLParser==0.0.2
nltk==3.6.5
numpy==1.22.4
pandas==1.3.4
PyGithub==1.55
PyYAML==6.0
requests==2.26.0
river==0.11.1
scikit_learn==1.1.1
spacy==3.2.3
stanza==1.4.0
tqdm==4.62.3
